+++ src/hooks/wistia.js
@@ -1,0 +1,24 @@
+import { useEffect } from 'react';
+
+const Wistia = () => {
+  useEffect(() => {
+    const script1 = document.createElement("script");
+    const script2 = document.createElement("script");
+
+    script1.src = "https://fast.wistia.com/embed/medias/nnkn75an3f.jsonp";
+    script1.async = true;
+
+    script2.src = "https://fast.wistia.com/assets/external/E-v1.js";
+    script2.async = true;
+
+    document.body.appendChild(script1);
+    document.body.appendChild(script2);
+
+    return () => {
+      document.body.removeChild(script1);
+      document.body.removeChild(script2);
+    }
+  });
+};
+
+export default Wistia;
